# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos y Cart",
        "description": "Documentación de la API para gestionar productos y ventas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Obtener productos",
                "description": "Devuelve una lista de productos, con la opción de filtrar por precio, disponibilidad o EAN",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "parameters": [
                    {
                        "name": "price_min",
                        "in": "query",
                        "description": "Precio mínimo",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "price_max",
                        "in": "query",
                        "description": "Precio máximo",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "available",
                        "in": "query",
                        "description": "Disponibilidad del producto (1 = disponible, 0 = no disponible)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ean",
                        "in": "query",
                        "description": "Código EAN del producto",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "name",
                                        "price",
                                        "stock",
                                        "ean"
                                    ],
                                    "properties": {
                                        "name": {
                                            "description": "Nombre del producto",
                                            "type": "string"
                                        },
                                        "price": {
                                            "description": "Precio del producto",
                                            "type": "string"
                                        },
                                        "stock": {
                                            "description": "Cantidad disponible en stock",
                                            "type": "integer"
                                        },
                                        "ean": {
                                            "description": "Código EAN del producto",
                                            "type": "string"
                                        },
                                        "active": {
                                            "description": "Estado de disponibilidad (1=activo, 0=inactivo)",
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros de filtro inválidos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Crear producto",
                "description": "Crea un nuevo producto en el sistema",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock",
                                    "ean"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del producto",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Precio del producto",
                                        "type": "string"
                                    },
                                    "stock": {
                                        "description": "Cantidad disponible en stock",
                                        "type": "integer"
                                    },
                                    "ean": {
                                        "description": "Código EAN del producto",
                                        "type": "string"
                                    },
                                    "active": {
                                        "description": "Estado de disponibilidad (1=activo, 0=inactivo)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "name",
                                        "price",
                                        "stock",
                                        "ean"
                                    ],
                                    "properties": {
                                        "name": {
                                            "description": "Nombre del producto",
                                            "type": "string"
                                        },
                                        "price": {
                                            "description": "Precio del producto",
                                            "type": "string"
                                        },
                                        "stock": {
                                            "description": "Cantidad disponible en stock",
                                            "type": "integer"
                                        },
                                        "ean": {
                                            "description": "Código EAN del producto",
                                            "type": "string"
                                        },
                                        "active": {
                                            "description": "Estado de disponibilidad (1=activo, 0=inactivo)",
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al crear el producto"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Actualizar producto",
                "description": "Actualiza un producto existente en el sistema",
                "operationId": "1419d479f9340e30e67fe21a2ca525b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock",
                                    "ean"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del producto",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Precio del producto",
                                        "type": "string"
                                    },
                                    "stock": {
                                        "description": "Cantidad disponible en stock",
                                        "type": "integer"
                                    },
                                    "ean": {
                                        "description": "Código EAN del producto",
                                        "type": "string"
                                    },
                                    "active": {
                                        "description": "Estado de disponibilidad (1=activo, 0=inactivo)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "name",
                                        "price",
                                        "stock",
                                        "ean"
                                    ],
                                    "properties": {
                                        "name": {
                                            "description": "Nombre del producto",
                                            "type": "string"
                                        },
                                        "price": {
                                            "description": "Precio del producto",
                                            "type": "string"
                                        },
                                        "stock": {
                                            "description": "Cantidad disponible en stock",
                                            "type": "integer"
                                        },
                                        "ean": {
                                            "description": "Código EAN del producto",
                                            "type": "string"
                                        },
                                        "active": {
                                            "description": "Estado de disponibilidad (1=activo, 0=inactivo)",
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al actualizar el producto"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticación para acceder a la API",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}